### 2026-04-03 #10373 @editorialbot
[PRE REVIEW]: CWIPI: Coupling With Interpolation Parallel Interface
**Submitting author:** @EricQuemerais (<a href="http://orcid.org/0009-0007-2018-6358">Eric Quémerais</a>)
**Repository:** https://github.com/onera/cwipi
**Branch with paper.md** (empty if default branch): joss_cwipi-1.3.0
**Version:** cwipi-1.3.0
