### 2026-04-03 #10373 @danielskatz
👋 @EricQuemerais - thanks for your submission. However, the paper is missing many of the required sections. Please see https://joss.readthedocs.io/en/latest/paper.html  Please ping me when these sections have been added.
