### 2026-04-15 #10373 @EricQuemerais
Hello @danielskatz, thanks for the feedback. I've pushed a new version of paper.md with the missing sections and fixed the incorrect DOIs in the references.
