### 2026-04-24 #10373 @MakisH
> [@MakisH](https://github.com/MakisH) could review [#10372](https://github.com/openjournals/joss-reviews/issues/10372) instead?

Discussing that in https://github.com/openjournals/joss-reviews/issues/10372#issuecomment-4315499463 Note that this review would also need some help/clarification regarding ONERA-internal resources/documentation.

Some notes from a first look through the paper:

- Interesting to see the layered approach, also in different JOSS papers (ParaDiGM -> CWIPI (-> OpenPALM)). I would like to see more details on this relation either in this paper, or in some public documentation. I always wanted to understand how the different couplers and their layers compare, and how CWIPI and OpenPALM relate.
- The _Summary_ and the _State of the Field_ also need more detail, as all these projects understand "coupling" a bit differently (at a different layer, and addressing a bit of different primary needs/applications). For example, MUI has the notion of "sampling" (originating from coupling flows or molecular dynamics), and preCICE has the notion of "mapping" (originating from fluid-structure interaction). MUI provides communication and space/time interpolation algorithms, but asks the user to implement the coupling algorithm (such as IQN). preCICE provides these coupling algorithms, and added time interpolation in v3. To my understanding from this paper, CWIPI mainly provides the mapping. While in many loosely-coupled problems, space interpolation (mapping) might be enough, strongly-coupled problems require such coupling algorithms, which are either implemented by the coupler, or implemented by the user. I would expect such a detail, for example, in this section. Is maybe OpenPALM the comparable layer?
- You write: _"A key strength of CWIPI lies in its ability to adapt to solvers' own numerical methods -- finite volumes, finite elements, discontinuous Galerkin, or spectral methods -- and to leverage their internal interpolation capabilities."_. I read this as "not black-box coupling", which is an important design decision with advantages and disadvantages. The "black-box nature" of MpCCI can be understood in two ways here: closed-course, or no access to internal numerics required. The first is a drawback, the second is an advantage, but also requires significant implementation effort for the coupler, and has accuracy tradeoffs.

For a comparison among couplers, you might find this older report useful: https://epubs.stfc.ac.uk/work/52144823 (comes from an MUI background and has some gaps in some couplers, but it is the most extensive comparison I know of).

@junghans let me know if you would prefer that I move these points into the review issue (https://github.com/openjournals/joss-reviews/issues/10424). Note that I am coming from the preCICE perspective (one of the other two couplers mentioned).

In any case, it will be very nice to eventually have a JOSS paper to cite for CWIPI!
