### 2026-04-28 #10424 @MakisH
- [x] 🟢 **Repository:** Is the source code for this software available at the [https://github.com/onera/cwipi](https://github.com/onera/cwipi)? -> Yes
- [x] 🟢 **License:** Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? -> Yes, LGPL3
- [x] 🟢 **Contribution and authorship:** Has the submitting author (@EricQuemerais) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? -> Yes: The submitting author has several commits and further maintainer activity over a long period of time. **TODO:** I have not yet checked the list of authors (data from GitHub not loading at the moment)

- [x] 🟢  **Development timeline:** Does the project show evidence of sustained development over time (preferably months or years) rather than rapid, recent code generation? -> Yes, the tool has been developed over several years.
