### 2026-05-06 onera/cwipi#2 @MakisH
c_new_api and c_new_api_surf_P1P0_P0P1_dynamic tests fail
In the process of reviewing the JOSS submission of CWIPI (https://github.com/openjournals/joss-reviews/issues/10424, cc @EricQuemerais), I tried running the tests on my laptop (Ubuntu 22.04).

The two lsited tests fail. Is that expected?

```
cwipi/build [joss_cwipi-1.3.0]$ ./cwp_run
[001/053] - c_new_api: Start
[002/053] - c_new_api_c_poly_block: Start
[003/053] - c_new_api_cellface_set_block: Start
[004/053] - c_new_api_surf_P1P0_P0P1: Start
[005/053] - c_new_api_surf_P1P0_P0P1_dynamic: Start
[006/053] - c_new_api_surf_P1P0_P0P1_without_part: Start
[007/053] - c_new_api_global_data: Start
[001/053] - c_new_api: Finalize
[001/053] - c_new_api: Failed
[002/053] - c_new_api_c_poly_block: Finalize
    Maximum of Maximum Resident Memory: 18.79 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.76 Mb on rank 0
[002/053] - c_new_api_c_poly_block: Passed - .38 s
[003/053] - c_new_api_cellface_set_block: Finalize
    Maximum of Maximum Resident Memory: 19.01 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.38 Mb on rank 0
[003/053] - c_new_api_cellface_set_block: Passed - .38 s
[004/053] - c_new_api_surf_P1P0_P0P1: Finalize
    Maximum of Maximum Resident Memory: 20.43 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.24 Mb on rank 0
[004/053] - c_new_api_surf_P1P0_P0P1: Passed - .36 s
[005/053] - c_new_api_surf_P1P0_P0P1_dynamic: Finalize
[005/053] - c_new_api_surf_P1P0_P0P1_dynamic: Failed
[006/053] - c_new_api_surf_P1P0_P0P1_without_part: Finalize
    Maximum of Maximum Resident Memory: 20.74 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.51 Mb on rank 0
[006/053] - c_new_api_surf_P1P0_P0P1_without_part: Passed - .35 s
[008/053] - c_new_api_part_data: Start
[009/053] - c_new_api_part_data2: Start
[010/053] - c_new_api_part_data_two_exch: Start
[011/053] - c_new_api_surf_P0P0: Start
[012/053] - c_new_api_surf_P0P0_disjoint_comm: Start
[013/053] - c_new_api_surf_P0P0_without_part: Start
[014/053] - c_new_api_user_interp_fct: Start
[015/053] - c_new_api_reverse_knn: Start
[016/053] - c_new_api_linear_P1P0_P0P1: Start
[007/053] - c_new_api_global_data: Finalize
    Maximum of Maximum Resident Memory: 18.19 Mb on rank 3
    Minimum of Maximum Resident Memory: 17.98 Mb on rank 1
[007/053] - c_new_api_global_data: Passed - .61 s
[008/053] - c_new_api_part_data: Finalize
    Maximum of Maximum Resident Memory: 19.16 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.35 Mb on rank 3
[008/053] - c_new_api_part_data: Passed - .58 s
[009/053] - c_new_api_part_data2: Finalize
    Maximum of Maximum Resident Memory: 20.82 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.78 Mb on rank 0
[009/053] - c_new_api_part_data2: Passed - .41 s
[010/053] - c_new_api_part_data_two_exch: Finalize
    Maximum of Maximum Resident Memory: 20.05 Mb on rank 0
    Minimum of Maximum Resident Memory: 19.84 Mb on rank 1
[010/053] - c_new_api_part_data_two_exch: Passed - .38 s
[011/053] - c_new_api_surf_P0P0: Finalize
    Maximum of Maximum Resident Memory: 20.92 Mb on rank 0
    Minimum of Maximum Resident Memory: 20.76 Mb on rank 1
[011/053] - c_new_api_surf_P0P0: Passed - .40 s
[012/053] - c_new_api_surf_P0P0_disjoint_comm: Finalize
    Maximum of Maximum Resident Memory: 20.78 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.69 Mb on rank 0
[012/053] - c_new_api_surf_P0P0_disjoint_comm: Passed - .37 s
[013/053] - c_new_api_surf_P0P0_without_part: Finalize
    Maximum of Maximum Resident Memory: 20.79 Mb on rank 0
    Minimum of Maximum Resident Memory: 20.55 Mb on rank 2
[013/053] - c_new_api_surf_P0P0_without_part: Passed - .37 s
[014/053] - c_new_api_user_interp_fct: Finalize
    Maximum of Maximum Resident Memory: 20.42 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.28 Mb on rank 0
[014/053] - c_new_api_user_interp_fct: Passed - .38 s
[015/053] - c_new_api_reverse_knn: Finalize
    Maximum of Maximum Resident Memory: 21.30 Mb on rank 0
    Minimum of Maximum Resident Memory: 21.12 Mb on rank 1
[015/053] - c_new_api_reverse_knn: Passed - .39 s
[017/053] - c_new_api_multiple: Start
[018/053] - c_new_api_variable_mesh: Start
[019/053] - c_new_api_param: Start
[020/053] - c_new_api_empty_parts: Start
[021/053] - c_new_api_remesh: Start
[022/053] - c_new_to_old_api: Start
[023/053] - c_new_vs_old_vol_sendrecv: Start
[024/053] - c_new_to_old_api_surf_P1P1: Start
[025/053] - c_new_api_vol: Start
[026/053] - c_new_api_vol_from_cellvtx: Start
[027/053] - c_linear_coupling: Start
[016/053] - c_new_api_linear_P1P0_P0P1: Finalize
    Maximum of Maximum Resident Memory: 20.46 Mb on rank 0
    Minimum of Maximum Resident Memory: 20.43 Mb on rank 1
[016/053] - c_new_api_linear_P1P0_P0P1: Passed - .40 s
[017/053] - c_new_api_multiple: Finalize
    Maximum of Maximum Resident Memory: 20.56 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.40 Mb on rank 0
[017/053] - c_new_api_multiple: Passed - .35 s
[018/053] - c_new_api_variable_mesh: Finalize
    Maximum of Maximum Resident Memory: 21.29 Mb on rank 0
    Minimum of Maximum Resident Memory: 21.22 Mb on rank 1
[018/053] - c_new_api_variable_mesh: Passed - .34 s
[019/053] - c_new_api_param: Finalize
    Maximum of Maximum Resident Memory: 18.33 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.28 Mb on rank 0
[019/053] - c_new_api_param: Passed - .33 s
[020/053] - c_new_api_empty_parts: Finalize
    Maximum of Maximum Resident Memory: 20.59 Mb on rank 0
    Minimum of Maximum Resident Memory: 20.01 Mb on rank 2
[020/053] - c_new_api_empty_parts: Passed - .33 s
[021/053] - c_new_api_remesh: Finalize
    Maximum of Maximum Resident Memory: 22.56 Mb on rank 0
    Minimum of Maximum Resident Memory: 22.11 Mb on rank 1
[021/053] - c_new_api_remesh: Passed - .63 s
[022/053] - c_new_to_old_api: Finalize
    Maximum of Maximum Resident Memory: 18.21 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.04 Mb on rank 0
[022/053] - c_new_to_old_api: Passed - .35 s
[023/053] - c_new_vs_old_vol_sendrecv: Finalize
    Maximum of Maximum Resident Memory: 25.36 Mb on rank 1
    Minimum of Maximum Resident Memory: 25.25 Mb on rank 0
[023/053] - c_new_vs_old_vol_sendrecv: Passed - .53 s
[024/053] - c_new_to_old_api_surf_P1P1: Finalize
    Maximum of Maximum Resident Memory: 20.80 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.79 Mb on rank 0
[024/053] - c_new_to_old_api_surf_P1P1: Passed - .38 s
[025/053] - c_new_api_vol: Finalize
    Maximum of Maximum Resident Memory: 20.98 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.81 Mb on rank 0
[025/053] - c_new_api_vol: Passed - .36 s
[026/053] - c_new_api_vol_from_cellvtx: Finalize
    Maximum of Maximum Resident Memory: 20.84 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.66 Mb on rank 0
[026/053] - c_new_api_vol_from_cellvtx: Passed - .35 s
[028/053] - c_surf_location_triaP2: Start
[029/053] - c_surf_location_curvetriaP2: Start
[030/053] - c_surf_coupling_P1P1_micron_polygon: Start
[031/053] - c_surf_coupling_P1P1_polygon: Start
[032/053] - c_surf_coupling_P1P1_polygon2: Start
[033/053] - c_surf_coupling_P1P1_polygon_with_external_points: Start
[034/053] - c_surf_coupling_P1P1: Start
[035/053] - c_surf_coupling_P1P0_P0P1: Start
[036/053] - c_surf_coupling_P1P0_P0P1_savedisk: Start
[037/053] - c_surf_coupling_P1P0_P0P1_savememory: Start
[038/053] - c_surf_coupling_P1P1_asynchronous: Start
[039/053] - c_surf_coupling_user_target_points: Start
[027/053] - c_linear_coupling: Finalize
    Maximum of Maximum Resident Memory: 18.79 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.47 Mb on rank 0
[027/053] - c_linear_coupling: Passed - .38 s
[028/053] - c_surf_location_triaP2: Finalize
    Maximum of Maximum Resident Memory: 18.42 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.38 Mb on rank 0
[028/053] - c_surf_location_triaP2: Passed - .40 s
[029/053] - c_surf_location_curvetriaP2: Finalize
    Maximum of Maximum Resident Memory: 18.62 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.40 Mb on rank 0
[029/053] - c_surf_location_curvetriaP2: Passed - .40 s
[030/053] - c_surf_coupling_P1P1_micron_polygon: Finalize
    Maximum of Maximum Resident Memory: 19.13 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.99 Mb on rank 1
[030/053] - c_surf_coupling_P1P1_micron_polygon: Passed - .42 s
[031/053] - c_surf_coupling_P1P1_polygon: Finalize
    Maximum of Maximum Resident Memory: 19.03 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.91 Mb on rank 0
[031/053] - c_surf_coupling_P1P1_polygon: Passed - .44 s
[032/053] - c_surf_coupling_P1P1_polygon2: Finalize
    Maximum of Maximum Resident Memory: 19.15 Mb on rank 1
    Minimum of Maximum Resident Memory: 19.08 Mb on rank 0
[032/053] - c_surf_coupling_P1P1_polygon2: Passed - .43 s
[033/053] - c_surf_coupling_P1P1_polygon_with_external_points: Finalize
    Maximum of Maximum Resident Memory: 18.90 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.83 Mb on rank 1
[033/053] - c_surf_coupling_P1P1_polygon_with_external_points: Passed - .38 s
[034/053] - c_surf_coupling_P1P1: Finalize
    Maximum of Maximum Resident Memory: 18.41 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.32 Mb on rank 1
[034/053] - c_surf_coupling_P1P1: Passed - .32 s
[035/053] - c_surf_coupling_P1P0_P0P1: Finalize
    Maximum of Maximum Resident Memory: 18.73 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.56 Mb on rank 0
[035/053] - c_surf_coupling_P1P0_P0P1: Passed - .30 s
[036/053] - c_surf_coupling_P1P0_P0P1_savedisk: Finalize
    Maximum of Maximum Resident Memory: 19.47 Mb on rank 0
    Minimum of Maximum Resident Memory: 19.43 Mb on rank 1
[036/053] - c_surf_coupling_P1P0_P0P1_savedisk: Passed - .47 s
[037/053] - c_surf_coupling_P1P0_P0P1_savememory: Finalize
    Maximum of Maximum Resident Memory: 64.03 Mb on rank 1
    Minimum of Maximum Resident Memory: 59.51 Mb on rank 0
[037/053] - c_surf_coupling_P1P0_P0P1_savememory: Passed - 2.73 s
[038/053] - c_surf_coupling_P1P1_asynchronous: Finalize
    Maximum of Maximum Resident Memory: 18.57 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.50 Mb on rank 0
[038/053] - c_surf_coupling_P1P1_asynchronous: Passed - .33 s
[040/053] - c_surf_coupling_user_interpolation: Start
[041/053] - c_surf_coupling_vector_exchange: Start
[042/053] - c_surf_coupling_output_error: Start
[043/053] - c_surf_coupling_simple_location: Start
[044/053] - c_surf_without_partitioning_coupling: Start
[045/053] - c_surf_without_partitioning_coupling2: Start
[046/053] - c_surf_sequential_coupling: Start
[047/053] - c_vol_coupling_P1P1: Start
[048/053] - c_vol_coupling_P1P1_with_external_points: Start
[049/053] - c_vol_coupling_micro_polyhedra_P1P1: Start
[050/053] - c_vol_coupling_polyhedra_P1P1: Start
[051/053] - c_vol_coupling_polyhedra_P1P1_with_external_points: Start
[039/053] - c_surf_coupling_user_target_points: Finalize
    Maximum of Maximum Resident Memory: 18.53 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.51 Mb on rank 0
[039/053] - c_surf_coupling_user_target_points: Passed - .33 s
[040/053] - c_surf_coupling_user_interpolation: Finalize
    Maximum of Maximum Resident Memory: 18.64 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.57 Mb on rank 0
[040/053] - c_surf_coupling_user_interpolation: Passed - .33 s
[041/053] - c_surf_coupling_vector_exchange: Finalize
    Maximum of Maximum Resident Memory: 18.58 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.48 Mb on rank 0
[041/053] - c_surf_coupling_vector_exchange: Passed - .33 s
[042/053] - c_surf_coupling_output_error: Finalize
    Maximum of Maximum Resident Memory: 18.63 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.56 Mb on rank 0
[042/053] - c_surf_coupling_output_error: Passed - .31 s
[043/053] - c_surf_coupling_simple_location: Finalize
    Maximum of Maximum Resident Memory: 18.39 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.35 Mb on rank 1
[043/053] - c_surf_coupling_simple_location: Passed - .32 s
[044/053] - c_surf_without_partitioning_coupling: Finalize
    Maximum of Maximum Resident Memory: 18.81 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.74 Mb on rank 0
[044/053] - c_surf_without_partitioning_coupling: Passed - .32 s
[045/053] - c_surf_without_partitioning_coupling2: Finalize
    Maximum of Maximum Resident Memory: 18.56 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.53 Mb on rank 1
[045/053] - c_surf_without_partitioning_coupling2: Passed - .29 s
[046/053] - c_surf_sequential_coupling: Finalize
    Maximum of Maximum Resident Memory: 18.76 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.54 Mb on rank 0
[046/053] - c_surf_sequential_coupling: Passed - .31 s
[047/053] - c_vol_coupling_P1P1: Finalize
    Maximum of Maximum Resident Memory: 18.52 Mb on rank 1
    Minimum of Maximum Resident Memory: 18.51 Mb on rank 0
[047/053] - c_vol_coupling_P1P1: Passed - .35 s
[048/053] - c_vol_coupling_P1P1_with_external_points: Finalize
    Maximum of Maximum Resident Memory: 18.58 Mb on rank 0
    Minimum of Maximum Resident Memory: 18.53 Mb on rank 1
[048/053] - c_vol_coupling_P1P1_with_external_points: Passed - .35 s
[049/053] - c_vol_coupling_micro_polyhedra_P1P1: Finalize
    Maximum of Maximum Resident Memory: 21.92 Mb on rank 1
    Minimum of Maximum Resident Memory: 19.82 Mb on rank 0
[049/053] - c_vol_coupling_micro_polyhedra_P1P1: Passed - .96 s
[050/053] - c_vol_coupling_polyhedra_P1P1: Finalize
    Maximum of Maximum Resident Memory: 21.91 Mb on rank 1
    Minimum of Maximum Resident Memory: 19.85 Mb on rank 0
[050/053] - c_vol_coupling_polyhedra_P1P1: Passed - 1.19 s
[052/053] - c_new_api_polygon_sol: Start
[053/053] - c_new_api_polygon_ex: Start
[051/053] - c_vol_coupling_polyhedra_P1P1_with_external_points: Finalize
    Maximum of Maximum Resident Memory: 21.92 Mb on rank 1
    Minimum of Maximum Resident Memory: 19.99 Mb on rank 0
[051/053] - c_vol_coupling_polyhedra_P1P1_with_external_points: Passed - .86 s
[052/053] - c_new_api_polygon_sol: Finalize
    Maximum of Maximum Resident Memory: 20.47 Mb on rank 1
    Minimum of Maximum Resident Memory: 20.42 Mb on rank 0
[052/053] - c_new_api_polygon_sol: Passed - .28 s
[053/053] - c_new_api_polygon_ex: Finalize
    Maximum of Maximum Resident Memory: 15.36 Mb on rank 1
    Minimum of Maximum Resident Memory: 15.24 Mb on rank 0
[053/053] - c_new_api_polygon_ex: Passed - .27 s

      002 tests failed out of 053 :
          - c_new_api
          - c_new_api_surf_P1P0_P0P1_dynamic

      000 tests produce vtk or log files out of 053 :

      000 ignored tests of 053 :

    elapsed cwp_run = 12.34 s
```

Happy to provide any further information you might need.

I understand that the CI of this project is in an internal GitLab, so, as an external, I cannot see if that is also happening there.
