### 2026-05-06 onera/cwipi#2 @EricQuemerais
c_new_api and c_new_api_surf_P1P0_P0P1_dynamic require 10 and 6 MPI ranks, respectively. Your installation may limit the number of MPI ranks at launch. In the test directory, you should find the outputs in the c_new_api_d and c_new_api_surf_P1P0_P0P1_dynamic_d folders to confirm this.
