### 2026-05-06 onera/cwipi#2 @MakisH
This can definitely be the case, since I only have 4 physical / 8 logical cores.

The respective directories exist, but they are empty. Since other tests have one subdirectory per MPI rank, I assume that these tests fail before even they can allocate the ranks, hence producing no output.

This requirement could be something to document in the [README](https://github.com/onera/cwipi/blob/master/README.md) for now, or even somewhere in the output of `cwp_run`.
