+++ docs/1.3.0/_sources/new_cwipi/concepts.rst.txt
@@ -282,11 +282,12 @@ For `Solver1` a field instance for sending the temperature will be created and a
 For `Solver2` the opposite will be done.
 
 There are ways to store field components:
-- CWP_FIELD_STORAGE_INTERLACED   : The number of components is constant for each element. The field is stored according to this pattern :math:` \left(c_{1,1} ... c_{s,1} ... c_{1,n} ... c_{s,n}\right)` , where :math:`s` is the number of components and :math:`n` the number of field elements ;
-- CWP_FIELD_STORAGE_INTERLEAVED  : The number of components is constant for each element. The field is stored according to this pattern :math:` \left(c_{1,1} ... c_{1,n} ... c_{s,1} ... c_{s,n}\right)` , where :math:`s` is the number of components and :math:`n` the number of field elements. In this mode,
+
+- CWP_FIELD_STORAGE_INTERLACED   : The number of components is constant for each element. The field is stored according to this pattern :math:`(c_{1,1} ... c_{s,1} ... c_{1,n} ... c_{s,n})`, where :math:`s` is the number of components and :math:`n` the number of field elements ;
+- CWP_FIELD_STORAGE_INTERLEAVED  : The number of components is constant for each element. The field is stored according to this pattern :math:`(c_{1,1} ... c_{1,n} ... c_{s,1} ... c_{s,n})`, where :math:`s` is the number of components and :math:`n` the number of field elements.
 
 
 Control Parameters
 ------------------
 
-*à remplir*
+*Not documented yet*
