+++ docs/1.3.0/breathe/doxygen/cwipi.cfg
@@ -4,7 +4,7 @@ GENERATE_LATEX   = NO
 GENERATE_MAN     = NO
 GENERATE_RTF     = NO
 CASE_SENSE_NAMES = NO
-INPUT            = /home/bandrieu/workspace/cwipi/cwipi/src/cwipi.h /home/bandrieu/workspace/cwipi/cwipi/src/new/cwp.h /home/bandrieu/workspace/cwipi/cwipi/src/new/client_server/client.h /home/bandrieu/workspace/cwipi/cwipi/tests/tutorial/c_new_api_polygon_sol.c /home/bandrieu/workspace/cwipi/cwipi/tests/tutorial/fortran_new_api_polygon_sol.F90 /home/bandrieu/workspace/cwipi/cwipi/tests/tutorial/python_new_api_polygon_sol.py /home/bandrieu/workspace/cwipi/cwipi/pattern/user_interp_function_c_pattern.txt /home/bandrieu/workspace/cwipi/cwipi/pattern/user_interp_function_fortran_pattern.txt /home/bandrieu/workspace/cwipi/cwipi/pattern/user_interp_function_python_pattern.txt
+INPUT            = /stck/thennion/cwipi/src/cwipi.h /stck/thennion/cwipi/src/new/cwp.h /stck/thennion/cwipi/src/new/client_server/client.h /stck/thennion/cwipi/tests/tutorial/c_new_api_polygon_sol.c /stck/thennion/cwipi/tests/tutorial/fortran_new_api_polygon_sol.F90 /stck/thennion/cwipi/tests/tutorial/python_new_api_polygon_sol.py /stck/thennion/cwipi/pattern/user_interp_function_c_pattern.txt /stck/thennion/cwipi/pattern/user_interp_function_fortran_pattern.txt /stck/thennion/cwipi/pattern/user_interp_function_python_pattern.txt
 ENABLE_PREPROCESSING = YES
 QUIET            = YES
 JAVADOC_AUTOBRIEF = YES
