### 2026-06-23 onera/cwipi#2 @ndelling
Hello @MakisH
A comment was added to the README on the `joss_cwipi-1.3.0` branch and a wrong default init in `cwp_run` was fixed. An explicit error is raised by `cwp_run` when the number of available ranks is not high enough to run the tests.
