+++ docs/1.0.0/_sources/new_cwipi/global_data.rst.txt
@@ -1,0 +1,6 @@
+.. _global data:
+
+Exhange of Global Data
+""""""""""""""""""""""
+
+*à remplir*
