+++ docs/1.0.0/_sources/new_cwipi/spatial_interp.rst.txt
@@ -1,0 +1,138 @@
+.. _spatial_interp:
+
+
+Since version 1.0, CWIPI offers multiple spatial interpolation methods.
+A specific spatial interpolation method can be affected to each Coupling object.
+
+Each method has a set of properties that can be adjusted (with ``CWP_Spatial_interp_property_set`` in C).
+Note that some method do not support all types of dof location.
+Finally, each method relies on some geometric mapping data to perform spatial interpolation.
+This data is computed by CWIPI and can be accessed to design :ref:`customized interpolation functions <Advanced functionalities>`.
+
+.. TODO: maybe mettre ce bloc warning à un autre endroit?
+
+.. warning::
+  Some spatial interpolation methods leave some target dofs unmapped if the interface meshes overlap partially.
+  The interpolated field for such targets is simply not computed.
+  Only the first :math:`n_{mapped} \times n_{comp}` values of the receiving data array are thus filled by CWIPI, where :math:`n_{mapped}` is the number of mapped targets (returned by ``CWP_N_computed_tgts_get``), and :math:`n_{comp}` the number of field components (specified when creating the field and returned by ``CWP_Field_n_components_get``).
+  Please refer to section :ref:`Advanced functionalities` for some examples of use.
+
+The spatial interpolation methods can be divided into the following four families.
+
+Interpolation from nearest neighbors
+------------------------------------
+
+This family includes two methods:
+
+  * ``CWP_SPATIAL_INTERP_FROM_NEAREST_SOURCES_LEAST_SQUARES``: each target dof is mapped to its ``n_neighbors`` nearest source dofs.
+
+  * ``CWP_SPATIAL_INTERP_FROM_NEAREST_TARGETS_LEAST_SQUARES``: each source dof is mapped to its ``n_neighbors`` nearest target dofs (note that some targets may be mapped to zero source).
+
+.. TODO: schéma?
+
+**Default interpolation**
+  Both methods perform a Least Square interpolation weighted by inverse squared distance (see `[Nealen 2004] <http://www.nealen.de/projects/mls/asapmls.pdf>`_ for more detail).
+
+**Properties**
+
+  The following properties can be adjusted by the user:
+
+    * ``n_neighbors``: number of neighbors to interpolate from (`integer`, default value: 4)
+
+    * ``polyfit_degree``: degree of polynomial fit (`integer`, default value: 1)
+
+**Acceptable degrees-of-freedom**
+
+  Since these are point cloud-based methods, they support all kinds of dof locations (mesh nodes, cell centers and user-defined points) for both source and target fields.
+
+
+Interpolation from mesh intersection
+------------------------------------
+
+This family includes only one method: ``CWP_SPATIAL_INTERP_FROM_INTERSECTION`` which consists on computing the overlap fractions between the elements of both interface meshes.
+It is suited for conservative interpolation.
+
+**Default interpolation**
+
+  The interpolated field at target :math:`T` is computed as follows
+
+  .. math::
+
+   f(T) = \frac{\sum_{S} w_S f(S)}{\sum_{S} w_S},
+
+  where :math:`S` designates a source element and :math:`w_S` is the volume (or area for surface coupling interfaces) of :math:`T \cap S`.
+
+  Note that for each mapped target element, the interpolation weights sum to one.
+  Target elements that intersect no source element are simply not mapped.
+
+
+**Properties**
+
+  .. _bbox paragraph:
+
+  The intersection algorithm uses efficient bounding box collisions detection as a first step.
+  The boxes can be inflated using a relative tolerance.
+  This is especially useful for non-planar, surface interface meshes where alignment with cartesian axes might cause some detection misses if the tolerance is set too low.
[…]
