+++ docs/dev/breathe/doxygen/cwipi.cfg
@@ -1,0 +1,17 @@
+PROJECT_NAME     = "cwipi"
+OUTPUT_DIRECTORY = cwipi
+GENERATE_LATEX   = NO
+GENERATE_MAN     = NO
+GENERATE_RTF     = NO
+CASE_SENSE_NAMES = NO
+INPUT            = /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/src/cwipi.h /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/src/new/cwp.h /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/src/new/client_server/client.h /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/tests/tutorial/c_new_api_polygon_sol.c /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/tests/tutorial/fortran_new_api_polygon_sol.F90 /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/tests/tutorial/python_new_api_polygon_sol.py /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/pattern/user_interp_function_c_pattern.txt /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/pattern/user_interp_function_fortran_pattern.txt /stck/ndelling/.jacamar-ci/builds/FOZNv1T6F/000/numerics/coupling/cwipi/pattern/user_interp_function_python_pattern.txt
+ENABLE_PREPROCESSING = YES
+QUIET            = YES
+JAVADOC_AUTOBRIEF = YES
+JAVADOC_AUTOBRIEF = NO
+GENERATE_HTML = NO
+GENERATE_XML = YES
+ALIASES = rst="\verbatim embed:rst"
+ALIASES += endrst="\endverbatim"
+ALIASES += inlinerst="\verbatim embed:rst:inline"
+

[…]
