### 2026-08-28 0062eb4
[cwp_run] fix NAvailRank for CI
