### 2026-09-01 7153ab4
[cwp_run] more compact fix
