+++ script/cwp_run.in
@@ -191,9 +191,7 @@ NrankUser=0
 runMake=0
 Test=""
 IsSrunMPI=0
-if [[ -z "${NAvailRank+x}" ]]; then
-  NAvailRank=$(getconf _NPROCESSORS_CONF)
-fi
+NAvailRank="${NAvailRank:-$(getconf _NPROCESSORS_CONF)}"
 failedTest=()
 testWithOutput=()
 testIgnored=()
