### 2026-09-01 fc3af4d
[ci] add intel25
