### 2026-09-04 63c4cac
[doc] Quickstart: Mention the source of the files
