### 2026-09-04 dc185f6
[doc] Quickstart: Add a visualization section
