+++ doc/sphinx/source/quick_start.rst
@@ -85,3 +85,17 @@ Execution
   export CWIPI_INSTALL_DIR=<path>/<where>/<cwipi>/<is>/<installed>
   export PYTHONPATH=$CWIPI_INSTALL_DIR/lib/python<version>/site-packages:$PYTHONPATH
   mpirun -np 2 python <file>.py
+
+Visualization
+~~~~~~~~~~~~~
+
+The code will produce a ``cwipi_writer`` directory, with the output of the two codes:
+
+- ``code1_code2_code1_code2`` (coupling ``code1_code2``, interface mesh for ``code1``, output by ``code2``)
+- ``code1_code2_code2_code1`` (coupling ``code1_code2``, interface mesh for ``code2``, output by ``code1``)
+
+Open the ``.case`` files with, e.g., Ensight or ParaView.
+The two fields ``s_a~super~fancy~field1`` (sent) and ``r_a~super~fancy~field1`` (received) show the same values on the two interface meshes.
+
+.. image:: ./images/quickstart-visualization.png
+   :scale: 50%
