### 2026-09-04 #10424 @MakisH
I have updated [my checklist](https://github.com/openjournals/joss-reviews/issues/10424#issuecomment-4333512718) to mark some issues as resolved, and I have added a new section on follow-up comments to the software paper.

I have also opened another PR with improvements to the Quickstart (https://github.com/onera/cwipi/pull/8), but still cannot run the Fortran and Python codes. The Fortran one looks outdated to me; the Python might be an issue on my end. @GuilloteauQ did you manage to run these two?
