### 2026-09-04 #10424 @EricQuemerais
> [@EricQuemerais](https://github.com/EricQuemerais) are you back yet?

Unless we've overlooked something, @GuilloteauQ 's latest comments have been addressed in the dev branch. I'll go through @MakisH's comments on the paper.
