### 2026-09-05 9e3c01d
[doc] Quickstart: Clarify that cwipi_configf.h is generated
