### 2026-09-07 9c99fd4
Fix typos in the documentation
