### 2026-09-09 3b5184c
[doc] Quickstart: Minor improvements
