### 2026-09-09 42e355e
Fix typos in the documentation
