+++ doc/sphinx/source/new_cwipi/global_data.rst
@@ -3,4 +3,4 @@
 Exchange of Global Data
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-*à remplir*
+*Not documented yet*
