### 2026-09-18 b284ea0
[cmake] fix for paradigm-2.8.1
