+++ paper/joss/paper.md
@@ -49,14 +49,13 @@ Existing coupling frameworks differ significantly in scope, geometry support, an
 
 Regarding temporal coupling algorithms, and unlike black-box solutions such as preCICE, CWIPI does not offer any turn-key solutions; instead, it solely provides the building blocks to define the orchestration of exchanges. Consequently, the temporal coupling algorithm is defined by the user in a decentralized manner.
 
-Regarding the spatial interpolation of fields, a concept referred to as "mapping" or "sampling" depending on the framework, CWIPI offers greater versatility than existing solutions. While preCICE’s "remapping" or MUI’s "sampling" mainly rely on point cloud-based interpolation, CWIPI supports a broad spectrum of interface representations. This includes linear, surface, and volumetric elements, polygons and polyhedra, as well as higher-order elements with curved edges and faces. To leverage these geometries, CWIPI provides three families of interpolation methods based on an underlying geometric algorithm: one based on point clouds (similar to preCICE and MUI), one based on the location of target points within the source mesh, and one based on the intersection of source and target meshes. For each of these, a default method is provided, complemented by a callback interface that allows users to implement the interpolation method best suited to their specific context. While MUI also allows users to write a custom callback, it is restricted to point clouds; in contrast, CWIPI extends this capability to all three geometric families. Within a CWIPI callback, parallelism management is completely transparent: the user works exclusively with local data without having to manage remote communications, except in the specific case where they invoke internal functions of their own code that require its intracommunicator. It is worth noting that the family based on target point location was the first to be implemented, which explains its privileged use in the vast majority of CWIPI applications.
+Regarding spatial interpolation of fields, a concept designated as "mapping" or "sampling" depending on the framework, CWIPI, due to its HPC nature, has oriented itself exclusively towards spatial interpolation methods based on local neighboring data. This contrasts notably with preCICE which, in addition to local interpolation methods, early on proposed a set of global RBF-type interpolation methods, involving all nodes of the interface mesh.
 
-Strategies for launching and communication differ significantly across frameworks. MUI and CWIPI favor an approach based on MPI MPMD (Multiple Program, Multiple Data), whereas preCICE opts for a decoupled approach where each code is launched via its own independent MPI command, with the connection between processes established a posteriori. Although CWIPI offers a TCP/IP communication mode similar to preCICE, it is rarely used, as the requirements of HPC users almost exclusively necessitate the use of MPI.
+CWIPI offers greater versatility than the solutions mentioned in terms of geometric representation of interfaces. While preCICE's "mapping" or MUI's "sampling" rely mainly on point cloud-based interpolation, CWIPI supports a broad spectrum of interface representations. This includes linear, surface and volumetric elements, polygons and polyhedra, as well as higher-order elements with curved edges and faces. To exploit these geometries, CWIPI provides three families of interpolation methods based on an underlying geometric algorithm: one based on point clouds (similar to preCICE and MUI), one based on locating target points within the source mesh, and one based on the intersection of source and target meshes. For each of these, a default method is provided, supplemented by a callback interface allowing users to implement the interpolation method best suited to their specific context. If MUI also allows users to write a custom callback, this is limited to point clouds; conversely, CWIPI extends this capability to the three geometric families. Within a CWIPI callback, parallelism management is entirely transparent: the user works exclusively with local data, without having to manage remote communications, except in the specific case where they invoke internal functions of their own code requiring their intracommunicator. It should be noted that the family based on locating target points was the first to be implemented, which explains its privileged use in the vast majority of CWIPI applications.
 
[…]
